requirements.txt
pillow
TkinterDnD2
patool